repairing tests
