+++ process/block/baseProcess_test.go
@@ -309,7 +309,6 @@ func TestBlockProcessor_CheckBlockValidity(t *testing.T) {
 		initStore(),
 		&mock.HasherMock{},
 		&mock.MarshalizerMock{},
-		&mock.TxProcessorMock{},
 		&mock.AccountsStub{},
 		mock.NewOneShardCoordinatorMock(),
 		&mock.ForkDetectorMock{},
@@ -317,6 +316,8 @@ func TestBlockProcessor_CheckBlockValidity(t *testing.T) {
 		createGenesisBlocks(mock.NewOneShardCoordinatorMock()),
 		true,
 		&mock.RequestHandlerMock{},
+		&mock.TransactionCoordinatorMock{},
+		&mock.Uint64ByteSliceConverterMock{},
 	)
 	blkc := createTestBlockchain()
 	body := &block.Body{}
@@ -382,7 +383,6 @@ func TestVerifyStateRoot_ShouldWork(t *testing.T) {
 		store,
 		&mock.HasherStub{},
 		&mock.MarshalizerMock{},
-		&mock.TxProcessorMock{},
 		accounts,
 		mock.NewOneShardCoordinatorMock(),
 		&mock.ForkDetectorMock{},
@@ -390,6 +390,8 @@ func TestVerifyStateRoot_ShouldWork(t *testing.T) {
 		createGenesisBlocks(mock.NewOneShardCoordinatorMock()),
 		true,
 		&mock.RequestHandlerMock{},
+		&mock.TransactionCoordinatorMock{},
+		&mock.Uint64ByteSliceConverterMock{},
 	)
 	assert.True(t, bp.VerifyStateRoot(rootHash))
 }
@@ -406,7 +408,6 @@ func TestBlockProcessor_computeHeaderHashMarshalizerFail1ShouldErr(t *testing.T)
 		initStore(),
 		&mock.HasherStub{},
 		marshalizer,
-		&mock.TxProcessorMock{},
 		&mock.AccountsStub{},
 		mock.NewOneShardCoordinatorMock(),
 		&mock.ForkDetectorMock{},
@@ -414,6 +415,8 @@ func TestBlockProcessor_computeHeaderHashMarshalizerFail1ShouldErr(t *testing.T)
 		createGenesisBlocks(mock.NewOneShardCoordinatorMock()),
 		true,
 		&mock.RequestHandlerMock{},
+		&mock.TransactionCoordinatorMock{},
+		&mock.Uint64ByteSliceConverterMock{},
 	)
 	hdr, txBlock := createTestHdrTxBlockBody()
 	expectedError := errors.New("marshalizer fail")
@@ -442,7 +445,6 @@ func TestBlockPorcessor_ComputeNewNoncePrevHashShouldWork(t *testing.T) {
 		initStore(),
 		hasher,
 		marshalizer,
-		&mock.TxProcessorMock{},
 		&mock.AccountsStub{},
 		mock.NewOneShardCoordinatorMock(),
 		&mock.ForkDetectorMock{},
@@ -450,6 +452,8 @@ func TestBlockPorcessor_ComputeNewNoncePrevHashShouldWork(t *testing.T) {
 		createGenesisBlocks(mock.NewOneShardCoordinatorMock()),
 		true,
 		&mock.RequestHandlerMock{},
+		&mock.TransactionCoordinatorMock{},
+		&mock.Uint64ByteSliceConverterMock{},
 	)
 	hdr, txBlock := createTestHdrTxBlockBody()
 	marshalizer.MarshalCalled = func(obj interface{}) (bytes []byte, e error) {
