+++ process/coordinator/process_test.go
@@ -1,1 +1,253 @@
 package coordinator
+
+/*
+//------- processMiniBlockComplete
+
+func TestShardProcessor_ProcessMiniBlockCompleteWithOkTxsShouldExecuteThemAndNotRevertAccntState(t *testing.T) {
+	t.Parallel()
+
+	hasher := mock.HasherMock{}
+	marshalizer := &mock.MarshalizerMock{}
+	dataPool := mock.NewPoolsHolderFake()
+
+	//we will have a miniblock that will have 3 tx hashes
+	//all txs will be in datapool and none of them will return err when processed
+	//so, tx processor will return nil on processing tx
+
+	txHash1 := []byte("tx hash 1")
+	txHash2 := []byte("tx hash 2")
+	txHash3 := []byte("tx hash 3")
+
+	senderShardId := uint32(1)
+	receiverShardId := uint32(2)
+
+	miniBlock := block.MiniBlock{
+		SenderShardID:   senderShardId,
+		ReceiverShardID: receiverShardId,
+		TxHashes:        [][]byte{txHash1, txHash2, txHash3},
+	}
+
+	tx1Nonce := uint64(45)
+	tx2Nonce := uint64(46)
+	tx3Nonce := uint64(47)
+
+	//put the existing tx inside datapool
+	cacheId := process.ShardCacherIdentifier(senderShardId, receiverShardId)
+	dataPool.Transactions().AddData(txHash1, &transaction.Transaction{
+		Nonce: tx1Nonce,
+		Data:  txHash1,
+	}, cacheId)
+	dataPool.Transactions().AddData(txHash2, &transaction.Transaction{
+		Nonce: tx2Nonce,
+		Data:  txHash2,
+	}, cacheId)
+	dataPool.Transactions().AddData(txHash3, &transaction.Transaction{
+		Nonce: tx3Nonce,
+		Data:  txHash3,
+	}, cacheId)
+
+	tx1ExecutionResult := uint64(0)
+	tx2ExecutionResult := uint64(0)
+	tx3ExecutionResult := uint64(0)
+
+	bp, _ := blproc.NewShardProcessor(
+		&mock.ServiceContainerMock{},
+		dataPool,
+		initStore(),
+		hasher,
+		marshalizer,
+		&mock.AccountsStub{
+			RevertToSnapshotCalled: func(snapshot int) error {
+				assert.Fail(t, "revert should have not been called")
+				return nil
+			},
+			JournalLenCalled: func() int {
+				return 0
+			},
+		},
+		mock.NewMultiShardsCoordinatorMock(3),
+		&mock.ForkDetectorMock{},
+		&mock.BlocksTrackerMock{},
+		createGenesisBlocks(mock.NewMultiShardsCoordinatorMock(3)),
+		true,
+		&mock.RequestHandlerMock{},
+		&mock.TransactionCoordinatorMock{},
+		&mock.Uint64ByteSliceConverterMock{},
+	)
+
+	err := bp.ProcessMiniBlockComplete(&miniBlock, 0, func() bool {
+		return true
+	})
+
+	assert.Nil(t, err)
+	assert.Equal(t, tx1Nonce, tx1ExecutionResult)
+	assert.Equal(t, tx2Nonce, tx2ExecutionResult)
+	assert.Equal(t, tx3Nonce, tx3ExecutionResult)
+}
+
+func TestShardProcessor_ProcessMiniBlockCompleteWithErrorWhileProcessShouldCallRevertAccntState(t *testing.T) {
+	t.Parallel()
+
+	hasher := mock.HasherMock{}
+	marshalizer := &mock.MarshalizerMock{}
+	dataPool := mock.NewPoolsHolderFake()
+
+	//we will have a miniblock that will have 3 tx hashes
+	//all txs will be in datapool and none of them will return err when processed
+	//so, tx processor will return nil on processing tx
+
+	txHash1 := []byte("tx hash 1")
+	txHash2 := []byte("tx hash 2 - this will cause the tx processor to err")
+	txHash3 := []byte("tx hash 3")
+
+	senderShardId := uint32(1)
+	receiverShardId := uint32(2)
+
+	miniBlock := block.MiniBlock{
+		SenderShardID:   senderShardId,
+		ReceiverShardID: receiverShardId,
+		TxHashes:        [][]byte{txHash1, txHash2, txHash3},
+	}
+
+	tx1Nonce := uint64(45)
+	tx2Nonce := uint64(46)
+	tx3Nonce := uint64(47)
+
+	errTxProcessor := errors.New("tx processor failing")
+
+	//put the existing tx inside datapool
+	cacheId := process.ShardCacherIdentifier(senderShardId, receiverShardId)
+	dataPool.Transactions().AddData(txHash1, &transaction.Transaction{
+		Nonce: tx1Nonce,
+		Data:  txHash1,
+	}, cacheId)
+	dataPool.Transactions().AddData(txHash2, &transaction.Transaction{
+		Nonce: tx2Nonce,
+		Data:  txHash2,
+	}, cacheId)
+	dataPool.Transactions().AddData(txHash3, &transaction.Transaction{
+		Nonce: tx3Nonce,
+		Data:  txHash3,
+	}, cacheId)
+
+	currentJournalLen := 445
+	revertAccntStateCalled := false
+
+	bp, _ := blproc.NewShardProcessor(
+		&mock.ServiceContainerMock{},
+		dataPool,
+		initStore(),
+		hasher,
+		marshalizer,
+		&mock.TxProcessorMock{
+			ProcessTransactionCalled: func(transaction *transaction.Transaction, round uint32) error {
+				if bytes.Equal(transaction.Data, txHash2) {
+					return errTxProcessor
+				}
+
+				return nil
+			},
+		},
+		&mock.AccountsStub{
+			RevertToSnapshotCalled: func(snapshot int) error {
+				if snapshot == currentJournalLen {
+					revertAccntStateCalled = true
+				}
+
+				return nil
+			},
+			JournalLenCalled: func() int {
+				return currentJournalLen
+			},
+		},
+		mock.NewMultiShardsCoordinatorMock(3),
+		&mock.ForkDetectorMock{},
+		&mock.BlocksTrackerMock{},
+		createGenesisBlocks(mock.NewMultiShardsCoordinatorMock(3)),
+		true,
+		&mock.RequestHandlerMock{},
+	)
+
+	err := bp.ProcessMiniBlockComplete(&miniBlock, 0, func() bool { return true })
+
+	assert.Equal(t, errTxProcessor, err)
+	assert.True(t, revertAccntStateCalled)
+}
+
+//------- receivedMiniBlock
+
+func TestShardProcessor_ReceivedMiniBlockShouldRequestMissingTransactions(t *testing.T) {
+	t.Parallel()
+
+	hasher := mock.HasherMock{}
+	marshalizer := &mock.MarshalizerMock{}
+	dataPool := mock.NewPoolsHolderFake()
+
+	//we will have a miniblock that will have 3 tx hashes
+	//1 tx hash will be in cache
+	//2 will be requested on network
+
+	txHash1 := []byte("tx hash 1 found in cache")
+	txHash2 := []byte("tx hash 2")
+	txHash3 := []byte("tx hash 3")
+
+	senderShardId := uint32(1)
+	receiverShardId := uint32(2)
+
+	miniBlock := block.MiniBlock{
+		SenderShardID:   senderShardId,
+		ReceiverShardID: receiverShardId,
+		TxHashes:        [][]byte{txHash1, txHash2, txHash3},
+	}
+
+	//put this miniblock inside datapool
+	miniBlockHash := []byte("miniblock hash")
+	dataPool.MiniBlocks().Put(miniBlockHash, miniBlock)
+
+	//put the existing tx inside datapool
+	cacheId := process.ShardCacherIdentifier(senderShardId, receiverShardId)
+	dataPool.Transactions().AddData(txHash1, &transaction.Transaction{}, cacheId)
+
+	txHash1Requested := int32(0)
+	txHash2Requested := int32(0)
+	txHash3Requested := int32(0)
+
+	requestHandler := &mock.RequestHandlerMock{}
+	requestHandler.RequestTransactionHandlerCalled = func(destShardID uint32, txHashes [][]byte) {
+		if containsHash(txHashes, txHash1) {
+			atomic.AddInt32(&txHash1Requested, 1)
+		}
+		if containsHash(txHashes, txHash2) {
+			atomic.AddInt32(&txHash2Requested, 1)
+		}
+		if containsHash(txHashes, txHash3) {
+			atomic.AddInt32(&txHash3Requested, 1)
+		}
+	}
+
+	bp, _ := blproc.NewShardProcessor(
+		&mock.ServiceContainerMock{},
+		dataPool,
+		initStore(),
+		hasher,
+		marshalizer,
+		&mock.TxProcessorMock{},
+		initAccountsMock(),
+		mock.NewMultiShardsCoordinatorMock(3),
+		&mock.ForkDetectorMock{},
+		&mock.BlocksTrackerMock{},
+		createGenesisBlocks(mock.NewMultiShardsCoordinatorMock(3)),
+		true,
+		requestHandler,
+	)
+
+	bp.ReceivedMiniBlock(miniBlockHash)
+
+	//we have to wait to be sure txHash1Requested is not incremented by a late call
+	time.Sleep(time.Second)
+
+	assert.Equal(t, int32(0), atomic.LoadInt32(&txHash1Requested))
+	assert.Equal(t, int32(1), atomic.LoadInt32(&txHash2Requested))
+	assert.Equal(t, int32(1), atomic.LoadInt32(&txHash2Requested))
+}
+*/
